% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parse.R
\name{abbrev}
\alias{abbrev}
\title{Abbreviate multiple matches}
\usage{
abbrev(x)
}
\arguments{
\item{x}{A vector of highlighted match strings}
}
\value{
A concatenated version from <= 3 match elements
}
\description{
Concatenates with ' ... '
}

