% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_ecosdocs.R
\name{add_raw_txt}
\alias{add_raw_txt}
\title{Add raw text to a data.frame for loading to Elastic}
\usage{
add_raw_txt(df)
}
\arguments{
\item{df}{A data.frame from a \code{bulk_*_prep} function, with txt_path var}
}
\value{
df, with a raw_txt variable filled using readLines
}
\description{
Add raw text to a data.frame for loading to Elastic
}
\examples{
one or more lines to demo the function
}

