% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_ecosdocs.R
\name{add_raw_txt_2}
\alias{add_raw_txt_2}
\title{Add raw text in N-character chunks to df, with df expansion}
\usage{
add_raw_txt_2(df)
}
\arguments{
\item{df}{A data.frame from a \code{bulk_*_prep} function, with txt_path var}
}
\value{
df, with a raw_txt variable filled using readLines
}
\description{
We use text extracted from PDFs with \link[pdftext]{pdftext} for full-text
search with Elastic, added in a single field, \code{raw_txt}. The volume of
data in \code{raw_txt} can be rather large; we recommend that it is added
}
\examples{
one or more lines to demo the function
}

