% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_analyzers.R
\name{alt_esadocs_analyzer}
\alias{alt_esadocs_analyzer}
\title{Alt custom elasticsearch analyzer for ESAdocs}
\usage{
alt_esadocs_analyzer()
}
\value{
A custom elasticsearch analyzer as a string JSON
}
\description{
Building custom analyzers for elasticsearch is a bit of a pain: they either
need to be text JSON or nested lists. This will, hopefully, make it a little
easier to tweak a basic custom analyzer. This is a string JSON representation
of \link{esadocs_analyzer}
}
\examples{
alt_esadocs_analyzer()
}

