% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parse.R
\name{build_query}
\alias{build_query}
\title{Prepare a user's query for elasticsearch}
\usage{
build_query(q)
}
\arguments{
\item{q}{The input query}
}
\value{
Query formatted for query_string or simple_query_string
}
\description{
Prepare a user's query for elasticsearch
}

