% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_ecosdocs.R
\name{bulk_ecosdocs_prep}
\alias{bulk_ecosdocs_prep}
\title{description 1-liner}
\usage{
bulk_ecosdocs_prep(docs, dates)
}
\arguments{
\item{paramName}{explanation}
}
\value{
what is returned
}
\description{
More detailed description of the function
}
\examples{
one or more lines to demo the function
}
\seealso{
if any see alsos
}

