% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_ecosdocs.R
\name{bulk_recplan_prep}
\alias{bulk_recplan_prep}
\title{Prepare  recovery plan data for loading to Elastic}
\usage{
bulk_recplan_prep(docs, dates)
}
\arguments{
\item{docs}{A data.frame of basic doc title, link, species information}

\item{dates}{A data.frame similar to \code{docs}, with doc dates}
}
\value{
what is returned
}
\description{
Prepare  recovery plan data for loading to Elastic
}
\examples{
#one or more lines to demo the function
}
\seealso{
\link{bulk_fedreg_prep} \link{bulk_fiveyr_prep}
 \link{bulk_consultation_prep}
}

