% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_ecosdocs.R
\name{chunked_es_loading}
\alias{chunked_es_loading}
\title{Add ESAdocs to Elastic in 100-doc chunks}
\usage{
chunked_es_loading(df, index = "esadocs", type)
}
\arguments{
\item{df}{A docs data.frame from \code{bulk_*_prep}}

\item{index}{The index into which documents are loaded}

\item{type}{The document type in df}
}
\value{
Nothing; chunking messages are printed during processing
}
\description{
The raw text of ESA docs can be rather large; rather than trying to hold
all of the text for all docs in a single data.frame, then load the data.frame
to Elastic in one fell swoop, we take a chunking approach. The raw text is
read in to the data.frame with \link{add_raw_txt} for each doc in each
100-doc chunk, then that chunk is loaded using \link[elastic]{docs_bulk}
}
\examples{
\dontrun{
chunked_es_loading(fr_dat, "esadocs", "federal_register")
}
}

