% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_mappings.R
\name{consult_mapping}
\alias{consult_mapping}
\title{Elasticsearch document mapping for ESA consultation docs}
\usage{
consult_mapping()
}
\arguments{
\item{...}{Additional properties to add to the mapping (not enabled yet)}
}
\value{
A document mapping as a nested list.
}
\description{
Custom mapping of consultation document variables in ESAdocs search. Most
variables remain \code{not_analyzed} to facilitate filtering.
}
\examples{
consult_mapping()
}

