% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_analyzers.R
\name{esadocs_analyzer}
\alias{esadocs_analyzer}
\title{Custom elasticsearch analyzer for ESAdocs}
\usage{
esadocs_analyzer(tokenizer = "standard", filter = list("standard",
  "lowercase", "stop", "cust_shingle"), min_shingle = "2",
  max_shingle = "3")
}
\arguments{
\item{tokenizer}{The type of tokenizer to use ['standard']}

\item{filter}{A list of \href{token filters}{https://www.elastic.co/guide/en/elasticsearch/reference/current/analysis-tokenfilters.html} for the analyzer ['standard',
'lowercase', 'stop', 'cust_shingle']}

\item{min_shingle}{The minimum word n-gram length ['2']}

\item{max_shingle}{The maximum word n-gram length ['3']}
}
\value{
A custom elasticsearch analyzer as a nested list
}
\description{
Building custom analyzers for elasticsearch is a bit of a pain: they either
need to be text JSON or nested lists. This will, hopefully, make it a little
easier to tweak a basic custom analyzer.
}
\examples{
esadocs_analyzer()
}

