% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_mappings.R
\name{fed_reg_mapping}
\alias{fed_reg_mapping}
\title{Elasticsearch document mapping for Federal Register documents}
\usage{
fed_reg_mapping()
}
\arguments{
\item{...}{Additional properties to add to the mapping (not enabled yet)}
}
\value{
A document mapping as a nested list.
}
\description{
Custom mapping of Federal Register variables in ESAdocs search. Most
variables remain \code{not_analyzed} to facilitate filtering.
}
\examples{
fed_reg_mapping()
}

