% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_mappings.R
\name{fiveyr_mapping}
\alias{fiveyr_mapping}
\title{Elasticsearch document mapping for ESA five-year reviews}
\usage{
fiveyr_mapping()
}
\arguments{
\item{...}{Additional properties to add to the mapping (not enabled yet)}
}
\value{
A document mapping as a nested list.
}
\description{
Custom mapping of five-year review variables in ESAdocs search. Most
variables remain \code{not_analyzed} to facilitate filtering.
}
\examples{
fiveyr_mapping()
}

