% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_gathering.R
\name{gather_pdfs}
\alias{gather_pdfs}
\title{A helper function to find relevant PDFs in one directory and copy to another}
\usage{
gather_pdfs(basepath, target)
}
\arguments{
\item{basepath}{A character string of the root path to be searched}

\item{target}{A character string of the path to which PDFs are copied}
}
\value{
The number of files found and copied
}
\description{
A helper function to find relevant PDFs in one directory and copy to another
}
\examples{
\dontrun{
gather_pdfs("~/pdf_dir", "~/new_pdf_dir")
}
}

