% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_parse.R
\name{get_highlight}
\alias{get_highlight}
\title{Return highlighted text from elasticsearch results}
\usage{
get_highlight(res)
}
\arguments{
\item{res}{An Elastic result object}
}
\value{
A vector of highlighted matches
}
\description{
Highlighting the matched text can be used in displaying the search
results, much as The Google does.
}

