% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_doc_text}
\alias{load_doc_text}
\title{Return the raw text of a file or NA if file missing}
\usage{
load_doc_text(path)
}
\arguments{
\item{path}{The system path to the text file of interest}
}
\value{
A single string of the contents of the text file
}
\description{
Return the raw text of a file or NA if file missing
}
\examples{
\dontrun{
load_doc_text("path/to/file.txt")
}
}

