% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_settings.R
\name{load_es_json}
\alias{load_es_json}
\title{Load a JSON for elastic settings}
\usage{
load_es_json(file)
}
\arguments{
\item{file}{Path to the JSON file to be loaded}
}
\value{
The contents of \code{file} as a list
}
\description{
Creating custom settings, e.g., analyzers and mappings, for elastic is a
bit tedious. The \code{body} variable in many \code{elastic} package functions
can be either a nested list or a JSON. Getting the formatting correct using
lists has been challenging for all but the simplest scenarios, and we decided
to use JSON for storing settings. But that has its own problems, e.g.,
combining different mappings in a single settings call requires a round-about
treatment of the JSON files. This function is the first step to loading
settings JSONs, but can't be further wrapped because of the need to combine
multiple JSONs for a single settings call.
}
\examples{
load_es_json("data-raw/recovery_plan_mapping")
}
\seealso{
\link{make_es_settings}
}

