% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{make_DTM}
\alias{make_DTM}
\title{Create the 'master' document-term matrix (DTM)}
\usage{
make_DTM(txt_dir)
}
\arguments{
\item{txt_dir}{Path to the directory containing all the txt files for the DTM}
}
\value{
Nothing. The DTM is written as an .rda file
}
\description{
More detailed description of the function
}
\examples{
make_DTM("~/ecos_docs/")
}
\seealso{
if any see alsos
}

