% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_file_paths}
\alias{make_file_paths}
\title{The base file name and corresponding text file name}
\usage{
make_file_paths(x)
}
\arguments{
\item{x}{The (string) file path}
}
\value{
A list with the pdf and txt file basenames
}
\description{
The base file name and corresponding text file name
}
\examples{
make_file_paths("this/is/a/file.pdf")
}

