% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{make_json}
\alias{make_json}
\title{Make a JSON of ESA doc files for elasticsearch}
\usage{
make_json(txt_dir, pdf_dir)
}
\arguments{
\item{txt_dir}{Path to the directory containing TXT files}
}
\value{
what is returned
}
\description{
More detailed description of the function
}
\examples{
\dontrun{
make_json("~/txts", "~/pdfs")
}
}
\seealso{
if any see alsos
}

