% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_mappings.R
\name{rec_plan_mapping}
\alias{rec_plan_mapping}
\title{Elasticsearch document mapping for ESA recovery plans}
\usage{
rec_plan_mapping()
}
\arguments{
\item{...}{Additional properties to add to the mapping (not enabled yet)}
}
\value{
A document mapping as a nested list.
}
\description{
Custom mapping of recovery plan variables in ESAdocs search. Most
variables remain \code{not_analyzed} to facilitate filtering.
}
\examples{
rec_plan_mapping()
}

