% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_parse.R
\name{result_asdf}
\alias{result_asdf}
\title{Extract data.frame from elasticsearch query result}
\usage{
result_asdf(res)
}
\arguments{
\item{res}{An elasticsearch query result}
}
\value{
A data.frame of document results
}
\description{
Elasticsearch query results are returned as a complex nested list that is
reasonably well-structured but a pain to parse. The basic \code{Search}
function of \link[elastic]{elastic} has a flag, \code{asdf}, to return the
results as a data.frame, but there is no generic function to extract the
data.frame. This function is tailored to the esadocs2 index, used in testing,
which includes a field (Scientific_Name) that is an array in elasticsearch
(list in R) and doesn't coerce to a single variable just using
\code{as.data.frame}.
}

