% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_nulls}
\alias{test_nulls}
\title{Test for NULL/NA/logical classes where these classes aren't expected}
\usage{
test_nulls(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
Logical; TRUE if NULL/NA/logical, else FALSE
}
\description{
Test for NULL/NA/logical classes where these classes aren't expected
}
\examples{
\dontrun{
tesf <- ("path/to/file.txt")
}
}

