% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_settings.R
\name{make_es_settings}
\alias{make_es_settings}
\title{Make elasticsearch index and search settings object}
\usage{
make_es_settings(analyzers = list(), mappings = list())
}
\arguments{
\item{analyzers}{A list of zero or more analyzers to include in settings}

\item{mappings}{A list of one or more document mappings to include}
}
\value{
A list of elasticsearch settings
}
\description{
<Need to add more...>
}
\examples{
make_es_settings(analyzers = list(esadocs_analyzer()),
                 mappings = list(fed_ref_mapping(), rec_plan_mapping()))
}
\seealso{
\link{esadocs_analyzer} \link{fed_reg_mapping}
}

